Generate async code
Disable it for the parameters flattening method, to avoid a binary
breaking change: it would add an async counterpart to an interface
method.
Moreover it generates an async counterpart due to a second level cache
check, which would incurs IO only if a distributed cache is enabled.
There is also an `IType.GetPropertyValue` call which has an async
counterpart, but it is called only on component types which currently
have a fully synchronous implementation for it.